backend/requirements.txt
alpaca-py==0.43.2
annotated-types==0.7.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
Flask==3.1.3
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==2.0.2
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
sseclient-py==1.9.0
ta==0.11.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.8
zipp==3.23.0
